fix: resolve x32 build warnings and Unity 64-bit assertion issues
- test_serialization.c: Cast dsdl_serialized_footprint() to size_t for x32 compatibility
- test_main.c: Replace UINT64 assertions with size_t assertions for Unity x32 support
- All tests now pass on both x64 and x32 (21/21 tests)

Coverage: 77.9% lines (target 90% - needs additional tests for remaining gaps)

+++ tests/test_serialization.c
@@ -1240,7 +1240,7 @@ static void test_serialize_exact_buffer_size(void)
     TEST_ASSERT_NOT_NULL(simple);
 
     // Get footprint (should be 7 bytes for 49 bits)
-    size_t footprint = dsdl_serialized_footprint(simple);
+    size_t footprint = (size_t)dsdl_serialized_footprint(simple);
     TEST_ASSERT_EQUAL_size_t(7, footprint);
 
     // Create field values
